Добавлены обработчики ошибок, связанных с неккоректным вводом данных.

+++ task-5-main-program-for-hw/bookstore_system/ui/controller/BookController.java
@@ -25,7 +25,7 @@ public class BookController {
         return reportFacade.getUnsoldBooks(sortByUnsoldBook);
     }
 
-    public String getBookDescription(int bookId) {
+    public String getBookDescription(Long bookId) {
         return reportFacade.getBookDescription(bookId);
     }
